\name{DistributionUtils-internal}
\alias{incompleteBesselKFor}

\title{Internal Distribution Utilities Functions}
\description{
  Internal functions for the package DistributionUtils
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.

  \code{incompleteBesselKFor} is a version of the incomplete Bessel
  function which calls a Fortran routine.
}
\keyword{ internal }
