\name{DistributionUtils-internal}

\title{Internal Distribution Utilities Functions}
\description{
  Internal functions for the package DistributionUtils
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
