\name{DistributionUtils-package}
\Rdversion{1.1}
\alias{DistributionUtils-package}
\alias{DistributionUtils}
\docType{package}
\title{
  Utility functions useful for all distributions in packages following
  the standard approach developed in Scott, Wuertz and Dong.
}
\description{
  The main functions calculate skewness and kurtosis, change the
  centering of moments and calculate moments by integration. There is
  also a general function for checking code for distributions.
}
\details{
\tabular{ll}{
Package: \tab DistributionUtils\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2010-01-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
David Scott <d.scott@auckland.ac.nz>

Maintainer: David Scott <d.scott@auckland.ac.nz>

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
\code{\link[fBasics:fBasics-package]{fBasics}} 
}
\examples{
is.wholenumber(-3:5)                           # TRUE
}
