\name{DistributionUtils-package}
\Rdversion{1.1}
\alias{DistributionUtils-package}
\alias{DistributionUtils}
\docType{package}
\title{
  Utility functions useful for all distributions in packages following
  the standard approach developed in Scott, Wuertz and Dong.
}
\description{
  The main functions calculate skewness and kurtosis, change the
  centering of moments and calculate moments by integration. There are a
  number of functions for checking code for distributions.
}
\details{
\tabular{ll}{
Package: \tab DistributionUtils\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2010-01-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Contains functions which are useful for packages implementing
distributions. Designed to work with my packages GeneralizedHyperbolic,
VarianceGamma and SkewHyperbolic.
}
\author{
David Scott <d.scott@auckland.ac.nz>

Maintainer: David Scott <d.scott@auckland.ac.nz>

}
\references{ 
  Scott, David J. and \enc{Würtz}{Wuertz}, Diethelm and Dong, Christine
  (2009) Software for Distributions in R. UseR: The R User
  Conference 2009
  \url{http://www.agrocampus-ouest.fr/math/useR-2009/slides/Scott+Wuertz+Dong.pdf}		  
}

\keyword{ package }
\seealso{
\code{\link[GeneralizedHyperbolic]{GeneralizedHyperbolicDistribution}} 
}
